requirements.txt
python-telegram-bot==13.3
pydantic==2.7.4
fastapi==0.111.0